repackage LogMDC
