+++ modules/activiti-engine/src/test/java/org/activiti/engine/test/logging/mdc/TestService.java
@@ -1,0 +1,32 @@
+package org.activiti.engine.test.logging.mdc;
+
+import org.activiti.engine.delegate.DelegateExecution;
+import org.activiti.engine.delegate.JavaDelegate;
+
+public class TestService implements JavaDelegate{
+	static String processInstanceId = null;
+	static String processDefinitionId = null;
+	static String executionId = null;
+	static String businessKey = null;
+	
+
+	@Override
+	public void execute(DelegateExecution execution) throws Exception {
+		TestService.processDefinitionId = execution.getProcessDefinitionId();
+		TestService.processInstanceId =  execution.getProcessInstanceId();
+		TestService.executionId = execution.getId();
+		TestService.businessKey = execution.getBusinessKey();
+		
+		Exception a = new Exception("test");
+		throw a;
+		
+	}
+	
+	public void clearProcessVariables() {
+		processDefinitionId = null;
+		processInstanceId = null;
+		executionId = null;
+		businessKey = null;
+	}
+
+}
